server/requirements.txt
flask
paho-mqtt
mysql-connector-python
requests
